refactor: DetailView를 사용하지 않고 뷰컨트롤러에서 뷰를 구현하도록 변경
- 새로운 할 일을 뷰모델이서 생성하도록 수정
